[FLINK-5229] [tm] Discard completed checkpoints if a subsequent operator's checkpoint fails
In case of a failure of any StreamOperator#snapshotState method, all up to this point created
StreamTaskStates are discarded. This ensures that a failing checkpoint operation of a chained
operator won't leave orphaned checkpoint data behind.

This closes #2924.

+++ flink-contrib/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/RocksDBStateBackend.java
@@ -77,6 +77,7 @@ import org.rocksdb.RocksDB;
 import org.rocksdb.RocksDBException;
 import org.rocksdb.RocksIterator;
 
+import org.rocksdb.Snapshot;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -638,10 +639,14 @@ public class RocksDBStateBackend extends AbstractStateBackend {
 	 */
 	private static class SemiAsyncSnapshot extends AsynchronousKvStateSnapshot<Object, Object, ValueState<Object>, ValueStateDescriptor<Object>, RocksDBStateBackend> {
 		private static final long serialVersionUID = 1L;
+
+		private final SerializableObject lock = new SerializableObject();
 		private final File localBackupPath;
 		private final URI backupUri;
 		private final List<StateDescriptor> stateDescriptors;
 		private final long checkpointId;
+
+		private volatile boolean discarded;
 
 		private SemiAsyncSnapshot(File localBackupPath,
 				URI backupUri,
@@ -651,22 +656,45 @@ public class RocksDBStateBackend extends AbstractStateBackend {
 			this.backupUri = backupUri;
 			this.stateDescriptors = columnFamilies;
 			this.checkpointId = checkpointId;
+			this.discarded = false;
 		}
 
 		@Override
 		public KvStateSnapshot<Object, Object, ValueState<Object>, ValueStateDescriptor<Object>, RocksDBStateBackend> materialize() throws Exception {
-			try {
-				long startTime = System.currentTimeMillis();
-				HDFSCopyFromLocal.copyFromLocal(localBackupPath, backupUri);
-				long endTime = System.currentTimeMillis();
-				LOG.info("RocksDB materialization from " + localBackupPath + " to " + backupUri + " (asynchronous part) took " + (endTime - startTime) + " ms.");
-				return new FinalSemiAsyncSnapshot(backupUri, checkpointId, stateDescriptors);
-			} catch (Exception e) {
-				FileSystem fs = FileSystem.get(backupUri, HadoopFileSystem.getHadoopConfiguration());
-				fs.delete(new org.apache.hadoop.fs.Path(backupUri), true);
-				throw e;
-			} finally {
-				FileUtils.deleteQuietly(localBackupPath);
+			synchronized (lock) {
+				if (discarded) {
+					throw new Exception("The SemiAsyncSnapshot has already been discarded.");
+				} else {
+					try {
+						long startTime = System.currentTimeMillis();
+						HDFSCopyFromLocal.copyFromLocal(localBackupPath, backupUri);
+						long endTime = System.currentTimeMillis();
+
+						LOG.info("RocksDB materialization from {} to {} (asynchronous part) took {} ms.", localBackupPath, backupUri, (endTime - startTime));
+
+						return new FinalSemiAsyncSnapshot(backupUri, checkpointId, stateDescriptors);
+					} catch (Exception e) {
+						FileSystem fs = FileSystem.get(backupUri, HadoopFileSystem.getHadoopConfiguration());
+						fs.delete(new org.apache.hadoop.fs.Path(backupUri), true);
+						throw e;
+					} finally {
+						discardState();
+					}
+				}
+			}
+		}
+
+		@Override
+		public void discardState() throws Exception {
+			if (!discarded) {
+				synchronized (lock) {
+					if (!discarded) {
+						discarded = true;
+						if (!FileUtils.deleteQuietly(localBackupPath)) {
+							LOG.warn("Could not delete the local backup file stored at {}.", localBackupPath);
+						}
+					}
+				}
 			}
 		}
 	}
@@ -732,9 +760,12 @@ public class RocksDBStateBackend extends AbstractStateBackend {
 		private transient org.rocksdb.Snapshot snapshot;
 		private transient AbstractStateBackend backend;
 
+		private final SerializableObject lock = new SerializableObject();
 		private final URI backupUri;
 		private final Map<String, Tuple2<ColumnFamilyHandle, StateDescriptor>> columnFamilies;
 		private final long checkpointId;
+
+		private volatile boolean discarded;
 
 		private FullyAsyncSnapshot(org.rocksdb.Snapshot snapshot,
 				AbstractStateBackend backend,
@@ -746,99 +777,122 @@ public class RocksDBStateBackend extends AbstractStateBackend {
 			this.backupUri = backupUri;
 			this.columnFamilies = columnFamilies;
 			this.checkpointId = checkpointId;
+			this.discarded = false;
 		}
 
 		@Override
 		public KvStateSnapshot<Object, Object, ValueState<Object>, ValueStateDescriptor<Object>, RocksDBStateBackend> materialize() throws Exception {
-			long startTime = System.currentTimeMillis();
-			CheckpointStateOutputView outputView;
+			synchronized (lock) {
+				if (discarded) {
+					throw new Exception("FullyAsyncSnapshot has already been discarded.");
+				} else {
+					long startTime = System.currentTimeMillis();
+					CheckpointStateOutputView outputView;
 
-			try {
-				try {
-					outputView = backend.createCheckpointStateOutputView(checkpointId, startTime);
-				} catch (Exception e) {
-					throw new Exception("Could not create a checkpoint state output view to " +
-						"materialize the checkpoint data into.", e);
-				}
-
-				try {
-					outputView.writeInt(columnFamilies.size());
-
-					// we don't know how many key/value pairs there are in each column family.
-					// We prefix every written element with a byte that signifies to which
-					// column family it belongs, this way we can restore the column families
-					byte count = 0;
-					Map<String, Byte> columnFamilyMapping = new HashMap<>();
-					for (Map.Entry<String, Tuple2<ColumnFamilyHandle, StateDescriptor>> column : columnFamilies.entrySet()) {
-						columnFamilyMapping.put(column.getKey(), count);
-
-						outputView.writeByte(count);
-
-						ObjectOutputStream ooOut = new ObjectOutputStream(outputView);
-						ooOut.writeObject(column.getValue().f1);
-						ooOut.flush();
-
-						count++;
-					}
-
-					ReadOptions readOptions = new ReadOptions();
-					readOptions.setSnapshot(snapshot);
-
-					for (Map.Entry<String, Tuple2<ColumnFamilyHandle, StateDescriptor>> column : columnFamilies.entrySet()) {
-						byte columnByte = columnFamilyMapping.get(column.getKey());
-
-						synchronized (dbCleanupLock) {
-							if (db == null) {
-								throw new RuntimeException("RocksDB instance was disposed. This happens " +
-									"when we are in the middle of a checkpoint and the job fails.");
-							}
-							RocksIterator iterator = db.newIterator(column.getValue().f0, readOptions);
-							iterator.seekToFirst();
-							while (iterator.isValid()) {
-								outputView.writeByte(columnByte);
-								BytePrimitiveArraySerializer.INSTANCE.serialize(iterator.key(),
-									outputView);
-								BytePrimitiveArraySerializer.INSTANCE.serialize(iterator.value(),
-									outputView);
-								iterator.next();
-							}
-						}
-					}
-				} catch (Exception e) {
 					try {
-						// closing the output view deletes the underlying data
-						outputView.close();
-					} catch (Exception closingException) {
-						LOG.warn("Could not close the checkpoint state output view. The " +
-							"written data might not be deleted.", closingException);
+						try {
+							outputView = backend.createCheckpointStateOutputView(checkpointId, startTime);
+						} catch (Exception e) {
+							throw new Exception("Could not create a checkpoint state output view to " +
+								"materialize the checkpoint data into.", e);
+						}
+
+						try {
+							outputView.writeInt(columnFamilies.size());
+
+							// we don't know how many key/value pairs there are in each column family.
+							// We prefix every written element with a byte that signifies to which
+							// column family it belongs, this way we can restore the column families
+							byte count = 0;
+							Map<String, Byte> columnFamilyMapping = new HashMap<>();
+							for (Map.Entry<String, Tuple2<ColumnFamilyHandle, StateDescriptor>> column : columnFamilies.entrySet()) {
+								columnFamilyMapping.put(column.getKey(), count);
+
+								outputView.writeByte(count);
+
+								ObjectOutputStream ooOut = new ObjectOutputStream(outputView);
+								ooOut.writeObject(column.getValue().f1);
+								ooOut.flush();
+
+								count++;
+							}
+
+							ReadOptions readOptions = new ReadOptions();
+							readOptions.setSnapshot(snapshot);
+
+							for (Map.Entry<String, Tuple2<ColumnFamilyHandle, StateDescriptor>> column : columnFamilies.entrySet()) {
+								byte columnByte = columnFamilyMapping.get(column.getKey());
+
+								synchronized (dbCleanupLock) {
+									if (db == null) {
+										throw new RuntimeException("RocksDB instance was disposed. This happens " +
+											"when we are in the middle of a checkpoint and the job fails.");
+									}
+									RocksIterator iterator = db.newIterator(column.getValue().f0, readOptions);
+									iterator.seekToFirst();
+									while (iterator.isValid()) {
+										outputView.writeByte(columnByte);
+										BytePrimitiveArraySerializer.INSTANCE.serialize(iterator.key(),
+											outputView);
+										BytePrimitiveArraySerializer.INSTANCE.serialize(iterator.value(),
+											outputView);
+										iterator.next();
+									}
+								}
+							}
+						} catch (Exception e) {
+							try {
+								// closing the output view deletes the underlying data
+								outputView.close();
+							} catch (Exception closingException) {
+								LOG.warn("Could not close the checkpoint state output view. The " +
+									"written data might not be deleted.", closingException);
+							}
+
+							throw new Exception("Could not write the checkpoint data into the checkpoint " +
+								"state output view.", e);
+						}
+					} finally {
+						discardState();
 					}
 
-					throw new Exception("Could not write the checkpoint data into the checkpoint " +
-						"state output view.", e);
-				}
-			} finally {
-				synchronized (dbCleanupLock) {
-					if (db != null) {
-						db.releaseSnapshot(snapshot);
+					StateHandle<DataInputView> stateHandle;
+
+					try {
+						stateHandle = outputView.closeAndGetHandle();
+					} catch (Exception ioE) {
+						throw new Exception("Could not close the checkpoint state output view and " +
+							"obtain the state handle.", ioE);
 					}
+
+					long endTime = System.currentTimeMillis();
+					LOG.info("Fully asynchronous RocksDB materialization to {} (asynchronous part) took {} ms.", backupUri, (endTime - startTime));
+					return new FinalFullyAsyncSnapshot(stateHandle, checkpointId);
 				}
-				snapshot = null;
 			}
-
-			StateHandle<DataInputView> stateHandle;
-
-			try {
-				stateHandle = outputView.closeAndGetHandle();
-			} catch (Exception ioE) {
-				throw new Exception("Could not close the checkpoint state output view and " +
-					"obtain the state handle.", ioE);
-			}
-
-			long endTime = System.currentTimeMillis();
-			LOG.info("Fully asynchronous RocksDB materialization to {} (asynchronous part) took {} ms.", backupUri, (endTime - startTime));
-			return new FinalFullyAsyncSnapshot(stateHandle, checkpointId);
 		}
 
+		@Override
+		public void discardState() throws Exception {
+			if (!discarded) {
+				final Snapshot snapshotToRelease = snapshot;
+
+				synchronized (lock) {
+					if (discarded) {
+						return;
+					} else {
+						discarded = true;
+						snapshot = null;
+					}
+				}
+
+				synchronized (dbCleanupLock) {
+					if (db != null) {
+						db.releaseSnapshot(snapshotToRelease);
+					}
+				}
+			}
+		}
 	}
 
 	/**
